Resolves #210 for me
Adding the `DROP TRIGGER IF EXISTS {trigger_name} ON {opts.db_table};` is harmless since the next step creates it again.  Not sure why this wasn't in here to begin with.
I think it might be needed for SQLite and MySQL but I don't have them setup to test with.